+++ app/controllers/admin/home/show_controller.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Admin::Home
+  class ShowController < ActionController::Base
+    include Authenticatable
+
+    layout "admin"
+  end
+end
