--- app/controllers/admin/home_controller.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class Admin::HomeController < ActionController::Base
-  include Authenticatable
-
-  layout "admin"
-end
