+++ app/controllers/welcome/show_controller.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Welcome
+  class ShowController < ActionController::Base
+    layout "application"
+  end
+end
