--- app/controllers/welcome_controller.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-class WelcomeController < ActionController::Base
-  layout "application"
-end
